+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-11-05
+
+### Added
+
+- More path variables(`%Downloads%`, `%Videos%`, `%Pictures%`, `%Music%`, `%Favorites%`)
+
+### Changed
+
+- Remove exclude software keyword
+- Optimize utils code
+
+### Fixed
+
+- Use the -d parameter to get the path name is incorrect.
+
 ## [2.0.0] - 2025-09-13
 
 ### Added